MBS-12155: Filter out downvoted tags from JSON WS user-tags/genres
These are already filtered out in _serialize_user_tag
and _serialize_user_genre in XMLSerializer,
but they were not being looked into in the JSON serializer.
This just adds an extra grep for is_upvote to
serialize_tags to effectively do the same there.

For some reason, we had no tests at all for JSON user-tags, too.
Added a JSON version of some of the ones we have
for XML, and modified the test data so that it will fail
if this bug is broken again.