MBS-12160: Use 'canonical' to display sorted edit data JSON
'canonical' ensures the JSON is sorted by key.
This makes things relatively slower, but it makes it easier
to compare different edits' data and to figure out
if something is missing or weird. Since the main target of the
edit data page are humans, it makes sense IMO to make
it easier for them to use by adding this consistency, even if
speed goes down a bit.