Enable Plack::Middleware::AccessLog in production
As noted in
https://github.com/metabrainz/syswiki/pull/45/files#r781430759, it's
often helpful to see the last request preceding an error or warning in
production.  Or the last request before a plackup process became stuck.

The AccessLog middleware is normally only enabled in development; if you
use plackup in development, you should be used to seeing requests
logged.  This enables the same logging in production.

This will increase the amount of logs generated by the website and web
service containers by quite a lot.  However, assuming we aggressively
fix spammy warnings and set the appropriate log-opts on the containers,
this patch would help make debugging alerts much easier.