MBS-2313: Avoid pointless empty annotation edits
This was only checking if latest_annotation existed
before happily using 'text', but 'text' is actually nullable.
It seems like emptying an annotation through the release editor
will not null it,
but emptying it through the usual edit annotation page will.
As such, the release editor will replace the null annotation
with an empty string one, in what is a completely pointless
exercise.
It might be preferable to just make annotation NOT NULL
and always ensure it's an empty string when blanked.