Add a default editor avatar for display
Removing Gravatar was great, but the voting icons
we use on edit notes expect an avatar to exist,
even if just a placeholder - without it, the display is a broken mess.
We could change the concept of how we show those,
but we do want to use (discourse) avatars in the future,
so it's probably best to not change it too much.
aerozol came up with a placeholder icon we can use;
for now it'll show for all editors, but we could still use it
for editors without an actual avatar once we have the discourse option.
I haven't made the same changes to the link_editor TT macro
because that's not breaking display and, honestly,
I didn't feel like figuring out how to import the image
on soon to be removed code.