Document Artist::Aliases test and remove cruft
The second half of the test here was checking whether
the page for an unrelated artist did not contain the alias,
which is absurd. It was also doing a basic JSON-LD test
of the index page of that artist, but we already do that
much better in Artist::Show and it's not connected with aliases.