Increase default avatar size & remove bg/border
The new placeholder editor.svg avatar is a bit hard to see at 12px, so
I'm increasing the default avatar size to 15px.

I found that it also looks better to display this without any white
background or gray border.  (I've kept these in cases where the avatar
is defined.)

The max-height rule has been removed because it conflicts with the fixed
size of the avatar.