Remove parallel processing on prove
We've been getting spurious failures on CircleCI which I have a feeling
are due to running certain tests in parallel; usually it's one of the
dump scripts under t/script/ that fail, but I'm not sure which tests are
conflicting with them.  Let's try disabling this and seeing if it
significantly affects the run-time.