Fix Data::Recording test flakiness
Very rarely, we'll get a failure like this:

        not ok 44 - recording 1 appears on エアリアル

        #   Failed test 'recording 1 appears on エアリアル'
        #   at t/lib/t/MusicBrainz/Server/Data/Recording.pm line 124.
        #          got: 'Brit Awards 2006'
        #     expected: 'エアリアル'

This is because Data::Recording::appears_on sorts RGs by their first
release date, and in this case there are *three* RG appearances with
NULL dates.  Normally it sorts エアリアル first because it's the first
one inserted, but that isn't any sort of guarantee.

Add an explicit release date on エアリアル to make sure it sorts before
the other NULL-dated RGs.