MBS-12183: Display previous value of annotation in AddAnnotation edits
We call everything "AddAnnotation", since they add
a new revision of the annotation, but there's quite a big difference
between cases where an annotation replaces another,
deletes it, or adds one where nothing existed.
This makes it so that a diff is shown when modifying
an existing annotation, similarly to the annotations-differences page.