Merge mostly useless Artist::Edits test and Artist::OpenEdits
Since no edits exist for the artist in this test, nothing is tested
except that the edit history page returns a 2XX.
It seems pointless to have a test just for that, so I merged it
into the more useful OpenEdits test and renamed that to Edits.
Also added a test to ensure that the edit does *not* appear
on the open edits list after accepting it.