Test user-facing warnings in Collection::Merge
We have a warning when merging public and private collections,
but we were not testing it.
Also, we were testing that a merge for different collection types
cannot be applied, but not that the user is blocked from entering it.
This tests that, and also that they're told what the problem is.