Document Collection::Show test and fix some issues
I'm pretty sure the xpath test for description was
hopelessly broken. It was checking for an id "collection" div
that does not seem to exist, and then checking whether
it had a specific element of class "deleted", but missing the quotes.
Luckily, it was also using not_ok, so it was actually checking
the element did *not* exist (although it should!),
and as such it was not failing.

A similar issue happened with the description check
in the no-description test.
@id=collection did not exist, so not_ok passed.
In fact, the description div *does* exist, it's just empty.