Autoflush script/server log output
Since timestamps have been removed of bash code and perl code in favor
of docker container (for servers) and ts (for cron tasks), logs should
be outputted as soon as possible to guarantee timestamp's accuracy.

This patch unbuffers STDOUT for the perl MusicBrainz::Server::Log module
that is consistenly used even in perl scripts.

Notes:
- Some external commands may still have buffered STDOUT;
- STDERR is unbuffered by default in perl.