MBS-12208: Withdrawn releases should count as official for overview
Withdrawn releases were by definition official until withdrawn,
so if a release group has only withdrawn releases, we should
probably still show it (this seems to be the majority community opinion,
anyway).
I understand two comparisons might be a lot more efficient
than any(1, 5), so doing that.

We need to run a one-off script to update RGs where a release was
*already* set to withdrawn, since otherwise these won't be set
to official(ish) until they are touched for some other reason.
Added the script to the upgrade file.