MBS-12213: Allow VGMDb artist links on places
VGMDb uses artist pages for studios, so we need to support them
on place entities. I added a new relationship type for it,
and now unblocking it in the code too.
There's a suggestion that org pages might also be used
for studios but I didn't find any, so not unblocking those just yet.