MBS-12195: Drop AcousticBrainz-related code (#2414)
We display some AcousticBrainz data on recording sidebars
and also a link on recording detail pages.
The project is being discontinued, although it is still going
to run until 2023. That said, given the reason to discontinue it
was specifically that the data isn't of the expected quality,
there seems to be no reason to keep showing it for a year
until it goes away.