Move 'CREATE TYPE' statements into CreateTypes.sql
These don't belong in CreateTables.sql; prior to this commit, you
weren't able to execute CreateTables.sql, DropTables.sql, and
CreateTables.sql again in that order, because the types weren't dropped
with the tables.  They still aren't, but now that they have their own
scripts, they're not *created* with the tables either.