Properly test text contents in details tab
This was testing content_contains, but that would still match
if there was a link to the MBID anywhere on the site, so the MBID
test was mostly useless. The last updated one was more useful
but would still have matched the sidebar date.
This now tests that the actual text (not just the html) contents,
and checks the label + content combo when useful to make sure
it's not matching something else on the page.
Will extend to other Details tests when documenting them
in the next commit.