Document and standardize Controller::Details tests
Added last_updated dates to all tested entities since it can't hurt
to be testing this.
The release test also checks sidebar content that isn't really
related to the Details page, but I thought I'd leave it be
because it doesn't hurt to test that the Details page also shows
the sidebar properly.