Extend useful Controller::Ratings tests to other entities
Except for the User test which is very different, all the entity
controller tests for Ratings but Artist literally only checked whether
the ratings page loaded. Now they actually check
whether the ratings page *works*, like the Artist one did.
I moved a bunch of the relevant SQL to a new controller_ratings.sql file
to avoid repeating it everywhere, but didn't go as far
as to create the entities in that file. Maybe we should, either now
or in a future revamp.