Add DataStore::RedisMulti
If the `DataStore::RedisMulti` module is active, then
`DATASTORE_REDIS_ARGS` may return an array ref of connection details.
(How do you know if it's active? grep for `DataStore::RedisMulti->new`.
We may revert back to plain-old `DataStore::Redis` if multiple instances
aren't needed.)

This module is useful when Redis service needs to be migrated to a new
server. We'll attempt to read from each connection in order (returning
the first non-empty result), and also distribute writes to all
connections. This allows time to copy any keys that don't exist on the
new instance from the old instance.