Return the original url if Amazon cleanup fails
This was returning an empty string if it couldn't properly
clean up the URL, but that makes it impossible to act on the
URL unless it exactly matches a cleanup,
and does not match other clean functions do.
As such, changed it to just return the URL, unchanged,
if nothing was done to it.