Move tests for the place index to Place::Show
There's no reason we should be testing that the area
containment is displayed on the main place page
as part of Place::Create. That's clearly a Place::Show test.
This commit moves it there, and by extension also
tests that area containment in JSON-LD works as expected
(which seems more useful than testing one level of areas).