MBS-12246: Also enable Genius artist links for places
For some reason, Genius lists places as artists.
We don't allow lyrics for places (because that makes no sense)
but in any case, the way they use them is to indicate where a song
was recorded and other similar, non-lyrics-related uses.
As such, "other databases" is a perfect fit for them.