MBS-12255: Add genre_alias_type table and make genre_alias consistent
For some reason (probably because it felt like overkill) we originally
implemented the genre_alias table without support
for alias types. That seems problematic though,
because there's at least the usual case of "typo" and
"name in a specific language", so "search hint" vs "genre name".
This adds types, and more generally changes
the genre_alias table to be consistent with the alias tables
for other entities.
To ensure the column order is also consistent,
we drop the old table and re-enter the old data
into a newly created one.
The genre_alias table should be empty in prod and any slaves,
but to be safe (in case some standalone users are using it)
this first copies any data to a temporary table and then
re-inserts it once done.