Use release_coverart_backup (#2455)
Columns in the release_coverart table have been getting cleared since
eb9637f469e18a236dd35f37651e22b48423731a was deployed.  I'm not sure
exactly how this is happening, but likely through the RebuildCoverArt
script, which I thought was disabled.

Unfortunately, this table is only in the private dump (where none is
available prior to 1/31), but I managed to find an older database dump
from May 2021 with a snapshot of the table.  For now I've imported that
into release_coverart_backup so that we can get the reports working
again.