Add AggregatedGenre, UserGenre for the web service
Having separate classes for these makes the code clearer, IMO, as there
were previously workarounds and indirection happening (see the changes
to _serialize_tags_and_ratings especially). We don't need the
tag-specific data -- which is really just the tag ID -- in these entity
genre lists, so it's more straightforward to construct genre-specific
classes that hold the data we need.

Loading the genre data directly onto these new classes of course means
we also don't have to call $c->model('Genre')->load on a list of tags
that are otherwise unneeded.

I've separated _serialize_entity_genre_list from _serialize_genre_list
so that we know the former operates on a list of aggregated genres,
while the latter just operates on a list of genres. (This isn't
necessary for tags because there's no such thing as a top-level tags
list.)