Order Data::Genre::get_all_limited by name
This affects the order of results for /ws/2/genre/all in the web service
(except for fmt=txt, which already returned all genre names in
alphabetical order).

This seems more logical to me because the pages themselves are ordered
by name. It's confusing to page the items by an invisible ID and then
jumble each pages' results by a different sort key; the sort key should
be consistent.