MBS-12258: Skip updating up-to-date containers
Skip updating containers which are running the latest built image for
this service already.  This change only helps with resuming an update,
by fastening further runs of the script after it has been interrupted
for some reason.

Implementation detail: `docker ps` shows the fully qualified name of the
image only if it matches the latest pulled image by this name, otherwise
it shows hash id instead.