MBS-12260: Block Facebook search URLs
There is no situation where this is a wanted link, so we might
as well block them. I made a generic, reusable error string
because I'm sure there are more of these.
Added an extra test to make sure we don't break legitimate links
that just start in the same way as a search results page.