Specify and test coordinate rounding
We were intentionally doing this already, but in such an unclear way
that it wasn't even obvious by looking at the code.
Added a comment to make it more obvious and a second test
that hits the use case to make it more clear there too
(the last test in this test file was already testing this in a way,
but its real purpose was a different one).