MBS-11904: Use all available space on "reorder relationships" edits (#2462)
Because we set the width of table.details headers to 240px by default,
this was getting unnecessarily compressed.
The relationship column contents are often very long,
while the other two columns are specifically set to be
as narrow as they can. As such, we should let the relationships column
take as much space as it needs. Added a "wide" class,
as a companion to the existing "narrow", because I expect
we might need this in a few other places at some point anyway.