Pass specific relationship ID as expected
This was actually buggy - the validation was not passing
the right id, so that the error would only be caught
if it didn't depend on the specific relationship
being expected. Since all existing errors were matched on URL
and seemingly did not depend on the link type itself,
we hadn't hit the bug before.