MBS-12279: Drop the manual label removal option
For some reason we still allow removing labels by hand,
even though they get auto-cleaned by ModBot when empty.
What's more, this removal can happen even if the label
has relationships, as long as it's not used as a release label.

It doesn't make sense to me at all to still allow manual removal
of labels. Most bad labels just get used as a release label
and become empty (auto-cleaned) as soon as the release is fixed.
In the rare case a label with relationships really doesn't exist
and should be removed, surely it's ok to remove the relationships
and let ModBot clean it up just like we would
for an artist in the same situation.
In fact, I found some comments from Ollie back during the
NGS migration era already saying that he expected labels
shouldn't be removable in the same ways already aren't,
so this isn't a new issue either...