Simplify passing user statistics
We want to reuse these in the admin user list,
and in any case it seems a lot simpler to follow
passing one stats block around rather than four separate ones.

Moved the vote stats calculation to Editor as well since there's
no real difference with the other ones.
Left the test as part of Data::Vote because it seems a lot simpler
to use the votes there for now than to recreate a full
voting history, but we might want to move it to Data::Editor
in the future, especially if we split Data::Vote into more tests.