MBS-11178: Make attribute lists visible to all users
This removes the info that is mostly useless for people without
editing rights (such as edit links) but displays everything else.

Row ids are not removed since documenting these on the site
is apparently helpful for people putting together SQL queries
(this has been requested).

Since this is visible to users, translations are brought back.