Generalize attribute layout
This reduces code reuse to some degree.
We could make further changes to also generalize
part of the tables for each model, but I'm not sure it's worth
the effort.
Additionally, this shows a proper translatable title
for each attribute type, since we'll display them to users now.
Only Language and Script were translated before.