MBS-12303: Display non-diff annotation data over all columns
There's no reason to stick all the data into the first column
when having a Diff element adds a second one.
The easiest solution seems to be to just add a colspan of 2;
even when the Diff is not present, all shows as normal
since everything then has the same colspan.