MBS-12302: Speed up large annotation diffing
Large annotation diffs like for edit #80101929 push the limits of
generic-diff, which isn't really designed to diff arrays of thousands of
characters.  It's so slow at doing this that edits like the one
referenced will simply error due to the template renderer socket timing
out and returning incomplete JSON.

The generic-diff README recommends fast-diff for efficient diffing of
text.  This is very fast as the name suggests, but doesn't allow for
custom "split" props, e.g. "\s+" for word diffing.  So we fall back to
fast-diff if the text to be diffed is greater than some arbitrary "large
string" length.  I picked 1024 characters.

It also makes sense to use fast-diff if the "split" prop is an empty
string, since we only perform character diffing in that case.