MBS-12305: Allow editing root instrument attribute
This allows displaying and editing the relationship attribute
"instrument" from the attributes list.
Since its children are automatically kept in sync with instruments,
editing links are hidden for them. We already returned 403
when trying to edit them; this keeps that, only changing it
for the root attribute which is not automatically edited itself.