MBS-12304: Only run guess feat on nonempty names (#2483)
It's pointless to try to guess feat artists from an empty name,
since it will obviously not contain any artist names.
This also avoids a warning when the name field is undefined,
rather than an empty string (e.g. when a blank release editor
is created and nobody has typed on the release name field).