MBS-12280 (I): Add target _blank to edit form links (#2481)
These are causing a content security issue that crashes the iframe
if the user eventually navigates outside mb.org
(e.g. to the forums). In any case, it seems unlikely the user
ever wants to load the documentation inside the iframe.

There's a separate link that needs the same changes
in the description of the instrument relationship attribute,
which cannot be edited from the site. As such,
marking this as part 1 of the ticket only.