MBS-12312: Convert history.tt to React
Rather than hiding the radio buttons when they can't be compared,
I chose to just disable them; seems cleaner to me to keep them onscreen.

TT did not care about whether there were even two different annotations
to compare (or any at all!) and happily loaded a semi-broken page.
I chose to just display a message if someone navigates to this page
by typing the URL in even though the entity has no annotation history,
and to actually show the history of 1 but not the bits related to
comparison if there's just one annotation in the history.