MBS-12323: Consider more person-only rels for ArtistsThatMayBePersons
Supporting musician is explicitly described as a person-to-artist rel.
Teacher is by definition person to person.
Artistic director and composer-in-residence are person-to-group
(maybe there's one or two edge cases of this not being the case,
but it should not be anywhere near common enough to cause problems).

By analogy, this also supports artistic director and
composer-in-residence for ArtistsThatMayBeGroups.