MBS-11599: Allow adding URL relationships to genre pages
Not implementing a full relationship editor yet since that
has no easy React implementation and mwiencek is working
on that for all entities, by which point it should be trivial
to also extend it to genres. In the meantime, being able to link
genres to Wikidata, Rate Your Music and whatnot seems like a great
first step.

We do need a "mock" relationship editor section for hidden inputs
to be attached to, though, so one is added.