MBS-12479: Add aliases for genres to the website
This just implements the basic ability to add genre aliases.
It does not yet do anything of more use with them than just
storing them (that is, it won't associate them with tags yet)
but it seems sensible to already allow storing the aliases
so that we have them already whenever we want to use them further.
Genre aliases will be added to the API with a later PR.

NB: GenreAliasType was generated with ./script/generate_code.pl