Block updates on relationship link tables
The link table and associated link_attribute tables mainly exist for
attribute caching: a lot of relationships happen to share the same link
type and attributes, so we can cache such data by a common ID and
minimize the amount of data that needs to be loaded across many
relationships.

Because they are used for caching and are shared across many (otherwise
unrelated) relationships, they are intended to be immutable: they should
only be inserted or deleted.

Updates to these tables should thus be blocked at the schema level.
This also helps ensure the data integrity of future materialized tables
that will rely on their immutability, like area_containment.