MBS-12328: Add #content to entity/Edits for artwork display
This got fixed with MBS-12294 by also checking for artwork
inside #edits but it turns out a bunch of pages have both
#content and #edits, leading to images being loaded into boxes
twice. On a quick check, it seems entity/Edits is the only
edit list missing #content anyway, for no clear reason,
so it seems a lot easier to just add a #content div to it
and revert the previous patch.