MBS-11920: Report for releases with low data quality
This allows users to easily find and fix low quality data for entities
in their subscriptions. It also makes it less stressful to just mark
something as low quality if an editor has no time or skill to fix it,
since now they will know other users (or even themselves
when they have the time) will be able to find it and hopefully
work on it.