MBS-12351: Also trim space only disambiguations
The transform was not being applied on spaces-only
disambiguation fields, because FormHandler considers it
empty. As such, I added validate_when_empty to enforce
always running the transform on the comment field.

This broke the test adding a rel, because we weren't entering
the comment field. That seemed weird to begin with (we are also
not changing name and sort name, yet we list them) so adding it here.