MBS-12350: Block smart links: bio.link
This also supports host-only links (e.g. https://blvtvntmusic.bio.link/)
so I added a mechanism to block the host as well for some shorteners.
We shouldn't do this for all of them since some do have
legitimate non-shortener uses.
That said, it's likely more of these should be blocked fully,
including the host, but I'm not going to do the research now, it can
be done whenever it is requested as needed.