MBS-12356: Correctly select + clean up Tidal store pages
The Tidal store is not a streaming page, but a download one.
We were automatically dropping the store bit and turning it
into a streaming link (not sure if the two always exist, but I wouldn't
be too confident, so we might have broken some links).
Since it's a legitimate download store, we should deal with it as such.
While Tidal has a country code for the store, it seems that the purchases
happen on the user's local store; removing the country code from a link
turns it into a redirect to the user's local store, so this just removes
the code.