MBS-8875: Improve CatNoLooksLikeASIN regexp
ASINs can be B plus any 9 letters or numbers,
but doing this creates a crazy amount of false positives
because it matches the catno layout of several labels,
including UMG. The best we can probably do is this:
B + 0 (still by far the most common) + 8 letters and digits,
of which at least one is a letter (since that still includes most ASINs
but excludes the UMG catnos).

Changed the catno in the release editor code to match
and added comments to keep them in sync.