Allow specifying the PG port for CheckSchemaMigration.sh
The version of PG I have installed for development is v14, but it's
important to test the schema migration against v12, since that's what we
use in production.  The easiest way to do this (for me) is to spin up a
postgres:12.10 Docker image running on another port.  The
CheckSchemaMigration script is modified here to support specifying a
custom port to connect to PG.

Note that any custom port must also be specified in DBDefs.pm for
MIGRATION_TEST1, MIGRATION_TEST2, and SYSTEM.

I haven't yet found a need to specify a custom host, but that would be
trivial to support in the future.