MBS-12042: Allow providing an edit note when editing annotation
Sometimes you want to provide a source for some data you add
to the annotation, or a detailed reason why you're changing something;
the changelog (which is supposed to be a short thing anyway)
does not seem too appropriate for this.
More generally, I do wonder whether we should drop the changelog
completely;
almost nobody uses it anyway AFAICT as anything else
than a place to put sources, which can now be put on the edit note.