Truncate tag_relation before deleting unused tags
Deleting unused tags from the tag table fails because tag_relation holds
foreign keys to them.

tag_relation is unused, non-replicated, and was apparently last
generated in 2014. There exists a script to re-calculate the table, when
we decide to use it again, at admin/CalculateRelatedTags.sh, but it
needs to be fixed first.