MBS-12380: Warn user about possible slow deletion
Deleting a user with many tags (10k or more) can be very slow
and even time out for the craziest amounts.
This lets the user know, and asks them to contact us
if the deletion fails.

We're hoping to eventually do this with a cron job instead,
so leaving a TODO to remove the warning if that happens.