Change source_type/target_type CoreEntityTypeT
This caused at least one property access on linkedEntities to return a
different type that includes null (since the type of the lookup key,
source_type, changed).  It seems plausible that an example might be
null, so I added a DeletedLink fallback for that case to handle the type
error.