MBS-12400: fix non-musicbrainz-schema dumps
Due to the `$table_noschema` addition in
0d3866005f7abdf4f1bc59d31ae5a6c4cba4a162, we were failing to include the
schema prefix in any dumped tables' file names.  Because MBImport.pl
expects the file names to match the table names as specified in
`@FULL_TABLE_LIST` (which includes schemas for non-musicbrainz-schema
tables), this causes the import to fail at finding data files for those
schemas.

The faulty change referenced above also has the side effect of causing
tables with the same names in different schemas to clobber each other in
the dumps.  For example, the cover-art-archive dump's art_type file
would contain rows from the event_art_archive.art_type table instead!

The only purpose of the `$table_noschema` change was to ensure that the
file names inside dbmirror v2 packets didn't include the "dbmirror2."
prefix.  It wouldn't cause any issue if they did, but it was preferred
to match how dbmirror v1 packets are dumped (whose tables are in the
musicbrainz schema, so are not schema-qualified).  I moved the
re-mapping from dump_table to ExportAllTables where it's needed.