MBS-12404: Replace Text::Unaccent (#2542)
As the ticket notes, Text::Unaccent is unmaintained and doesn't compile
against recent version of Perl [1].  I'm not sure since what version,
but the linked bug report is from 2018, and I can attest that I've been
having to patch Text::Unaccent manually on my development machines for
years.

I'm replacing it with Text::Unaccent::PurePerl, which as its name
suggests is written in Perl.  It passes all of our existing
`normalise_strings` tests, which contained the only reference to
Text::Unaccent.  We use `normalise_strings` mainly for two things:

 * Determining whether a title change is an auto-edit, to allow "simple"
   changes to go through automatically.

 * Determining whether a disambiguation comment is required when adding
   a new entity.

[1] https://rt.cpan.org/Public/Bug/Display.html?id=124815