MBS-12412: Fail gracefully if requesting an invalid ID in edit-cover-art
This would ISE if the user requested an ID that could not be found
by the first{} call, since we later tried to access $artwork->types.

This fails more gracefully now, like we already did when the release
had no artwork at all; I also added specific messages
for the 404 page for both cases.