Run all sql before enabling replication triggers (#2539)
I see no reason why replication triggers should be re-enabled *before*
the .standalone.sql file is run.  If the standalone file happens to
modify data in replicated tables, then the current order causes those
changes to be logged in the dbmirror pending tables, which is not ideal.
I've fixed the order so that replication is re-enabled only after all
migration SQL is run.