MBS-12336: Validate that Predicate::Set is being passed integers (#2509)
We always want an integer here (they're mostly rowIDs, except
for privileges and data quality which are also ints).
Default data quality is -1, so we can't restrict to *positive* integers.

Notice from the test that for some reason we semi-supported
passing the arguments in a comma-only list
('4,5' rather than two arguments 4 and 5). That *did* get
the intended results returned from the DB, but it breaks
display (no actual values are shown as selected in the search form)
and it seems like a huge hack, so I'm disallowing it with this
rather than special-casing it.