MBS-12418: Also format artists in setlists if there's no MBID link
I think having no "Artist:" nor <strong> for artist lines
unless there is an MBID link was actually not really intended,
and in any case it seems silly - it means that unless a link is added,
there's literally no point in adding @ vs * in the setlist.
This makes it so that artist lines are always displayed
starting with "Artist: " and in bold.