MBS-12415: Stop most interface wrapping
The main intention of the overflow-wrap changes was
to wrap absurdly long entity names, and very long words in edit notes.
By setting it everywhere, it was also affecting things like the
tag button on the sidebar, or barcodes in entity lists, which does not
seem useful. As such, this sets it only on <a> and on edit notes.