MBS-12419: Block Genius.com links at release level
People seem to add quite a few of this as discography entry,
or as a free download link if there is any link from the page
to a place where it can be downloaded (or, sometimes,
even if there isn't any).

Adding the error as a generic one since we might find more lyrics
sites that get used in this way.