MBS-12437: Limit overflow-wrap even more
Even after commit 4f47d6a6bc73d7f317b51386c8bf36dc7c7c49b7 limited
overflow-wrap to just <a> elements and edit notes, there are
still a bunch of things wrapping that probably should not,
and a lot of it seems to happen even when there's nothing that is
really long enough to require wrapping in the first place,
especially in tables with lots of columns.
As such, this commit reverts the generic overflow-wrap for <a>
and takes a bit-by-bit approach starting from the minimum
amount of wrapped things: entity headers (where anything too long
will break the interface for no good use), and URL entity links
(which can get really long), both on EntityLink and in the
Edit URL edit display.
The added wrapping for edit note texts hasn't caused complaints
so far, and tags on tag lists already had overflow-wrap for a long
time now without complaints, so leaving those in too.