MBS-12395: Report for videos in mediums that shouldn't support video
This will help find both videos that have been wrongly reused
in audio-only releases, and stuff like video CDs marked as just "CD".
I hand-picked a list of formats that do not support video
rather than a list of those that can because I'd rather have
missing entries if we add a new format that doesn't support video
than false positives if we add a new format that does.